moved view fns into view.rs
